# main/docs/swagger.yaml
basePath: /
definitions:
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  req.AddProductRemarkRequest:
    properties:
      order_product_id:
        description: 商品ID
        type: integer
      remark:
        description: 备注
        type: string
    type: object
...
  req.BuffetCustomerType:
    properties:
      customer_type_id:
        description: 自助餐顾客类型ID
        type: integer
      name_text:
        description: 自助餐顾客类型名称
        type: string
      num:
        description: 顾客数量
        type: integer
    type: object
...
  req.OpenDeskRequest:
    properties:
      buffet_customer_type:
        description: 自助餐顾客类型
        items:
          $ref: '#/definitions/req.BuffetCustomerType'
        type: array
      buffet_ids:
        description: 自助餐ID
        items:
          type: integer
        type: array
      is_buffet:
        description: 是否是餐桌
        type: integer
      meal_num:
        description: 顾客数量
        type: integer
    type: object
...
  req.PrintRechargeOrderReq:
    properties:
      print_lang:
        description: 打印语言
        type: string
      recharge_order_uuid:
        description: 充值订单Uuid
        type: integer
    required:
    - print_lang
    - recharge_order_uuid
    type: object
...
  req.RechargeOrderRefundReq:
    properties:
      refund_money:
        description: 部分退款金额
        type: number
      refund_type:
        description: '退款类型: 1-整单退款, 2-部分退款'
        enum:
        - 1
        - 2
        type: integer
      uuid:
        description: 充值订单uuid
        type: integer
    required:
    - refund_type
    - uuid
    type: object
  req.RechargeOrderUuidReq:
    properties:
      uuid:
        description: 充值订单uuid
        type: integer
    type: object
...
  resp.H5Buffet:
    properties:
      app_id:
        type: integer
      buffetCustomerType:
        items:
          $ref: '#/definitions/resp.H5BuffetCustomerType'
        type: array
      buffetTaxes:
        items:
          properties:
            app_id:
              type: integer
            buffet_id:
              type: integer
            buffet_tax_type:
              type: integer
            create_time:
              type: string
            id:
              type: integer
            tax_category_id:
              type: integer
            tax_rate:
              type: integer
            update_time:
              type: string
          type: object
        type: array
      buy_limit_status:
        type: integer
      create_time:
        type: string
      delete_time:
        type: integer
      id:
        type: integer
      is_comb:
        type: integer
      is_remain_continue:
        type: integer
      name:
        type: string
      name_text:
        type: string
      price:
        type: string
      remain_continue_notice_time:
        type: integer
      remain_continue_time:
        type: integer
      sale_num:
        type: integer
      shop_supplier_id:
        type: integer
      sort:
        type: integer
      status:
        type: integer
      time_limit:
        type: integer
      update_time:
        type: string
    type: object
  resp.H5BuffetCustomerType:
    properties:
      app_id:
        type: integer
      buffet_id:
        type: integer
      create_time:
        type: string
      customer_type_id:
        type: integer
      id:
        type: integer
      name_text:
        type: string
      price:
        type: string
      shop_supplier_id:
        type: integer
      update_time:
        type: string
    type: object
  resp.H5Category:
    properties:
      category_id:
        type: integer
      is_special:
        type: integer
      name:
        type: string
      name_text:
        type: string
      parent_id:
        type: integer
      path_name_text:
        type: string
      product_count:
        type: string
      product_num:
        type: integer
    type: object
  resp.H5CategoryList:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.H5Category'
        type: array
    type: object
...
  resp.H5Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  resp.PrinterLogData:
    properties:
      cashier_device_id:
        description: 收银机绑定的设备ID
        type: string
      copies:
        description: 打印机.份数 => 对应旧表print_times
        type: integer
      create_time:
        description: 日志创建时间戳
        type: integer
      data:
        description: 打印数据
        type: string
      data_type:
        description: 数据类型 1-预结账单 2-结账单 3-一菜一单 4-整单打印 5-打印发票 6-打印营业数据 7-打印交班单;
        type: integer
      first_execution:
        description: 是否首次执行打印 1-是 0-否;
        type: integer
      "no":
        description: 桌台号或者呼叫号(如果有)
        type: string
      num:
        description: 打印次数
        type: integer
      print_method:
        description: 打印方式 1文本打印, 2图片打印'
        type: integer
      printer_config:
        description: 打印机.配置
        type: string
      printer_time:
        description: 打印时间
        type: integer
      printer_type:
        description: 打印机.类型
        type: string
      printer_uuid:
        description: 打印机uuid
        type: integer
      reason:
        description: 原因
        type: string
      related_type:
        description: 关联类型
        type: integer
      related_uuid:
        description: 关联ID
        type: integer
      status:
        description: 状态(0结束,1进行中,2成功)
        type: integer
      type:
        description: 类型:0系统默认队列,1云上服务下放
        type: integer
      uuid:
        description: 打印日志Uuid
        type: integer
    type: object
...
  resp.RechargeOrderCashier:
    properties:
      real_name:
        description: 收银员真实姓名
        type: string
    type: object
  resp.RechargeOrderInfo:
    properties:
      amount:
        description: 实付金额：充值金额+手续费
        type: number
      cashier:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderCashier'
        description: 收银员
      charge_due:
        description: 找零
        type: number
      create_time:
        description: 下单时间
        type: integer
      extra:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderItemExtra'
        description: 通过当前数据控制按钮是否显示
      gift_amount:
        description: 赠送金额
        type: number
      gift_point:
        description: 赠送积分
        type: number
      member:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderMember'
        description: 会员信息
      operation_log:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderOperationLog'
        description: 操作日志
      order_no:
        description: 充值订单编号
        type: string
      payment_methods:
        description: 支付方式
        items:
          $ref: '#/definitions/resp.RechargeOrderPaymentMethod'
        type: array
      payment_time:
        description: 支付时间
        type: integer
      recharge_amount:
        description: 充值金额
        type: number
      status:
        description: 充值订单状态
        type: integer
      uuid:
        description: 充值订单Uuid
        type: integer
    type: object
  resp.RechargeOrderItem:
    properties:
      amount:
        description: 实付金额
        type: number
      cashier:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderCashier'
        description: 收银员
      extra:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderItemExtra'
        description: 通过当前数据控制按钮是否显示
      gift_amount:
        description: 赠送金额
        type: number
      gift_point:
        description: 赠送积分
        type: number
      member_uuid:
        description: 会员uuid
        type: integer
      order_no:
        description: 订单编号
        type: string
      payment_methods:
        description: 支付方式
        items:
          type: string
        type: array
      payment_time:
        description: 完成时间（支付时间）（时间戳）
        type: integer
      recharge_amount:
        description: 充值金额
        type: number
      refund_money:
        description: 退款金额
        type: number
      status:
        description: 订单状态 订单状态, 0-待付款、1-已完成、2-已取消
        type: integer
      uuid:
        description: 充值订单Uuid
        type: integer
    type: object
  resp.RechargeOrderItemExtra:
    properties:
      is_cell_cancel:
        description: 是否可取消
        type: boolean
      is_cell_refund:
        description: 是否可退款
        type: boolean
      is_cell_reverse_settle:
        description: 是否可反结账：同一个收银员工、同一个班次、已支付、未退款
        type: boolean
    type: object
  resp.RechargeOrderList:
    properties:
      list:
        description: 订单列表
        items:
          $ref: '#/definitions/resp.RechargeOrderItem'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/resp.RechargeOrderListMeta'
        description: Meta信息
    type: object
  resp.RechargeOrderListMeta:
    properties:
      cancel_num:
        description: 已取消数量
        type: integer
      complete_num:
        description: 已完成数量
        type: integer
      page_no:
        description: 当前页码
        type: integer
      page_size:
        description: 每页大小
        type: integer
      total:
        description: 总数
        type: integer
      unpaid_num:
        description: 待付款数量
        type: integer
    type: object
  resp.RechargeOrderMember:
    properties:
      nickname:
        description: 会员昵称
        type: string
      uuid:
        description: 会员Uuid
        type: integer
    type: object
  resp.RechargeOrderOperationLog:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.RechargeOrderOperationLogItem'
        type: array
    type: object
  resp.RechargeOrderOperationLogItem:
    properties:
      client:
        description: 来源
        type: string
      create_time:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      real_name:
        description: 操作员工真实姓名
        type: string
      username:
        description: 操作员工账号
        type: string
    type: object
  resp.RechargeOrderPaymentMethod:
    properties:
      code:
        description: 支付方式代号
        type: integer
      name:
        description: 支付方式名称
        type: string
      price:
        description: 支付金额
        type: number
      source_text:
        description: 来源
        type: string
    type: object
  resp.RechargeOrderRefundInfo:
    properties:
      gift_amount:
        description: 赠送金额
        type: number
      gift_point:
        description: 赠送积分
        type: number
      payment_records:
        description: 支付订单信息
        items:
          $ref: '#/definitions/resp.RefundRechargeOrderPaymentRecord'
        type: array
      recharge_amount:
        description: 充值金额
        type: number
      recharge_member_info:
        allOf:
        - $ref: '#/definitions/resp.RefundRechargeOrderMemberInfo'
        description: 充值会员信息
      refundable_amount:
        description: 可退款金额
        type: number
      uuid:
        description: 充值订单Uuid
        type: integer
    type: object
  resp.RechargeOrderReverseSettleInfo:
    properties:
      member_info:
        allOf:
        - $ref: '#/definitions/resp.ReverseSettleRechargeOrderMemberInfo'
        description: 会员信息
      message:
        description: 提示信息
        type: string
      status:
        description: 订单状态：0-正常；1-会员主账户余额不足；2-存在待支付订单
        type: integer
    type: object
  resp.RefundRechargeOrderMemberInfo:
    properties:
      balance:
        description: 主账户
        type: number
      gift_balance:
        description: 赠送账户
        type: number
      points:
        description: 积分
        type: number
    type: object
  resp.RefundRechargeOrderPaymentRecord:
    properties:
      payment_amount:
        description: 支付金额
        type: number
      payment_name:
        description: 支付方式名称
        type: string
      refundable_amount:
        description: 剩余可退款金额
        type: number
    type: object
...
  resp.ReverseSettleRechargeOrderMemberInfo:
    properties:
      balance:
        description: 主账户
        type: number
      gift_balance:
        description: 赠送账户
        type: number
      nickname:
        description: 会员昵称
        type: string
      points:
        description: 积分
        type: number
      uuid:
        description: 会员Uuid
        type: integer
    type: object
...
host: 127.0.0.1:8081
info:
  contact: {}
  description: 点餐系统服务端接口文档
  title: ttpos-server-go API
  version: "1.0"
paths:
...
  /cashier/recharge_order/cancel:
    post:
      consumes:
      - application/json
      description: 取消充值订单
      parameters:
      - description: 取消充值订单参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RechargeOrderUuidReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - JwtToken: []
      summary: 取消充值订单
      tags:
      - 收银端.充值订单
  /cashier/recharge_order/check_reverse_settle:
    get:
      consumes:
      - application/json
      description: 检查充值订单反结账信息
      parameters:
      - description: 充值订单Uuid
        in: query
        name: uuid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RechargeOrderReverseSettleInfo'
              type: object
      security:
      - JwtToken: []
      summary: 检查充值订单反结账信息
      tags:
      - 收银端.充值订单
  /cashier/recharge_order/info:
    get:
      consumes:
      - application/json
      description: 获取充值订单详情
      parameters:
      - description: 充值订单Uuid
        in: query
        name: uuid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RechargeOrderInfo'
              type: object
      security:
      - JwtToken: []
      summary: 获取充值订单详情
      tags:
      - 收银端.充值订单
  /cashier/recharge_order/list:
    get:
      consumes:
      - application/json
      description: 获取充值订单列表
      parameters:
      - description: 页码
        in: query
        name: page_no
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 订单编号
        in: query
        name: order_no
        type: string
      - description: 日期类型 -1=全都、 0=今天、 1=昨天、 2=本周
        in: query
        name: date_type
        type: integer
      - description: 启用添加时间 false-不启用，true-启用
        in: query
        name: enable_create_time
        type: boolean
      - description: 启用支付时间 false-不启用，true-启用
        in: query
        name: enable_payment_time
        type: boolean
      - description: 查询开始时间戳
        in: query
        name: query_start_time
        type: integer
      - description: 查询结束时间戳
        in: query
        name: query_end_time
        type: integer
      - description: 充值订单状态, -1=全都、 0=待付款、1=已完成、2=已取消
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RechargeOrderList'
              type: object
      security:
      - JwtToken: []
      summary: 获取充值订单列表
      tags:
      - 收银端.充值订单
  /cashier/recharge_order/print_ticket:
    post:
      consumes:
      - application/json
      description: 充值订单-打印小票
      parameters:
      - description: 充值订单打印小票参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.PrintRechargeOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.PrinterLogData'
              type: object
      security:
      - JwtToken: []
      summary: 充值订单-打印小票
      tags:
      - 收银端.充值订单
  /cashier/recharge_order/refund:
    get:
      consumes:
      - application/json
      description: 获取充值订单退款信息
      parameters:
      - description: 充值订单Uuid
        in: query
        name: uuid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.RechargeOrderRefundInfo'
              type: object
      security:
      - JwtToken: []
      summary: 获取充值订单退款信息
      tags:
      - 收银端.充值订单
    post:
      consumes:
      - application/json
      description: 充值订单退款
      parameters:
      - description: 退款充值订单参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RechargeOrderRefundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - JwtToken: []
      summary: 充值订单退款
      tags:
      - 收银端.充值订单
  /cashier/recharge_order/reverse_settle:
    post:
      consumes:
      - application/json
      description: 反结账充值订单
      parameters:
      - description: 反结账充值订单参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RechargeOrderUuidReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - JwtToken: []
      summary: 反结账充值订单
      tags:
      - 收银端.充值订单
...
  /h5/index.php/scan/base.base/getInfo:
    post:
      consumes:
      - application/json
      description: 获取桌码基础信息，整个h5应用需要的基础信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.GetBaseInfoResponse'
      security:
      - JwtToken: []
      summary: 桌码基础信息
      tags:
      - 扫码点餐
  /h5/index.php/scan/order.Order/buffetList:
    post:
      consumes:
      - application/json
      description: 获取自助餐套餐列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.H5Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.H5Buffet'
                  type: array
              type: object
      security:
      - JwtToken: []
      summary: 自助餐套餐信息
      tags:
      - 扫码点餐
  /h5/index.php/scan/order.Order/remark:
    post:
      consumes:
      - application/json
      description: 给商品添加备注
      parameters:
      - description: 添加备注参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.AddProductRemarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.H5Response'
      security:
      - JwtToken: []
      summary: 添加备注
      tags:
      - 扫码点餐
  /h5/index.php/scan/order.Order/setTable:
    post:
      consumes:
      - application/json
      description: 开台
      parameters:
      - description: 创建扫码桌台订单参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.OpenDeskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.H5Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.H5Buffet'
                  type: array
              type: object
      security:
      - JwtToken: []
      summary: 开台
      tags:
      - 扫码点餐
  /h5/index.php/scan/product.category/index:
    post:
      consumes:
      - application/json
      description: 获取商品分类
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.H5Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.H5CategoryList'
              type: object
      security:
      - JwtToken: []
      summary: 获取商品分类
      tags:
      - 扫码点餐
...
securityDefinitions:
  JwtToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"